permission parsing error fix

+++ meetings/permissions.py
@@ -39,7 +39,7 @@ class IsCustomerOwner(BasePermission):
         if not request.user.is_staff and not request.user.is_superuser:
             if request.user.id == (int)(view.kwargs["customer_pk"]):
                 has_permission = True
-                if view.action == 'create' and "customer" in request.data and (int)(view.kwargs["customer_pk"]) != request.data["customer"]:                 
+                if view.action == 'create' and "customer" in request.data and (int)(view.kwargs["customer_pk"]) != (int)(request.data["customer"]):                 
                     has_permission = False
                     self.message = "the meeting could not be created because the customer sent in URL does not match the one sent in data."                
                 if view.action == 'create' and "auditor" in request.data:
@@ -62,7 +62,7 @@ class IsAuditorOwner(BasePermission):
         if request.user.is_staff:
             if request.user.id == (int)(view.kwargs["auditor_pk"]):
                 has_permission = True
-                if view.action == "create" and "auditor" in request.data and not request.data["auditor"] == request.user.id:
+                if view.action == "create" and "auditor" in request.data and not (int)(request.data["auditor"]) == request.user.id:
                     has_permission = False
                     self.message = "Auditor sent in data is not the same auditor on url"
             elif request.user.is_superuser:
@@ -81,7 +81,7 @@ class CheckAuditor(BasePermission):
     
     def has_permission(self, request, view):
         has_permission = False
-        if "auditor" in request.data and request.data["auditor"] == request.user.id:
+        if "auditor" in request.data and (int)(request.data["auditor"]) == request.user.id:
             has_permission = True
         return has_permission
